Use -include-pkg-deps to also depend on system and package .hi files.
This way we can detect correctly when external code changes.

For a case where this failed before, see https://github.com/JohnLato/parmake-bug.

This also needs a change to `BuildPlan.new`:
When determining the initially buildable modules, before we chose the ones with exactly one dependency.
Now, we don't count external files into those dependencies, as they do not need to be built.